add-some-unit-tests Done refactoring.
