+++ graphjet-core/src/test/java/com/twitter/graphjet/algorithms/filter/ANDFilterTest.java
@@ -17,7 +17,11 @@
 package com.twitter.graphjet.algorithms.filter;
 
 import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.List;
 
+import com.twitter.graphjet.algorithms.filters.RecentTweetFilter;
+import com.twitter.graphjet.stats.StatsReceiver;
 import org.junit.Test;
 
 import com.twitter.graphjet.algorithms.filters.ANDFilters;
@@ -27,6 +31,7 @@ import com.twitter.graphjet.hashing.SmallArrayBasedLongToDoubleMap;
 import com.twitter.graphjet.stats.NullStatsReceiver;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 
 public class ANDFilterTest {
   // Dummy filter that always return true
@@ -83,5 +88,27 @@ public class ANDFilterTest {
     filters.add(trueFilter);
     andFilter = new ANDFilters(filters, new NullStatsReceiver());
     assertEquals(true, andFilter.filterResult(0, new SmallArrayBasedLongToDoubleMap[]{}));
+  }
+
+  @Test
+  public void testFilterResultReturningFalseOne() {
+    List<ResultFilter> linkedList = new LinkedList<ResultFilter>();
+    StatsReceiver statsReceiver = new NullStatsReceiver();
+    ANDFilters aNDFilters = new ANDFilters(linkedList, statsReceiver);
+    SmallArrayBasedLongToDoubleMap[] smallArrayBasedLongToDoubleMapArray = new SmallArrayBasedLongToDoubleMap[9];
+
+    assertFalse(aNDFilters.filterResult(1633L, smallArrayBasedLongToDoubleMapArray));
+  }
+
+  @Test
+  public void testFilterResultReturningFalseTwo() {
+    List<ResultFilter> linkedList = new LinkedList<ResultFilter>();
+    StatsReceiver statsReceiver = new NullStatsReceiver();
+    RecentTweetFilter recentTweetFilter = new RecentTweetFilter(2147483639L, statsReceiver);
+    linkedList.add(recentTweetFilter);
+    ANDFilters aNDFilters = new ANDFilters(linkedList, statsReceiver);
+    SmallArrayBasedLongToDoubleMap[] smallArrayBasedLongToDoubleMapArray = new SmallArrayBasedLongToDoubleMap[5];
+
+    assertFalse(aNDFilters.filterResult(1L, smallArrayBasedLongToDoubleMapArray));
   }
 }
